Fix pathing through caves and use packedWorldPoint instead of WorldPoint
Pathing through caves, e.g. south-east of Trollheim, was not working because the cave transports did not have a travel time which made the cost calculate based on distance. This was fixed by forcing all transport nodes to use travel time and not defaulting to distance, while letting non-transport nodes not include travel time.

Removed as many instances of WorldPoint as possible by instead using packed WorldPoint integers. Next up will be to implement an iterator to replace java.util.HashMap with PrimitiveIntHashMap