Wilderness Slayer Cave and Revenant Caves entrances/exits
Currently no check for Revenant Caves 100,000 coins entry fee, but this can probably be payed directly from the bank. Would then require tracking both whether it is already paid and whether the user has enough money in the bank, while also disregarding whether the money is in the inventory or not. Probably needs a `Remote` true/false column for item requirements.