Prepare for detecting built POH transports
The POH transports (e.g. Varrock Portal, Ornate jewellery box, ...) must be saved in the config to be able to distinguish between different in-game accounts. The format is a string containing ID and world location pairs like e.g. "33092=1928 5731 0;33433=1935 5731 0". Make sure that the world locations are template coordinates and not instance coordinates (not x bigger than 6000), but should take into account the custom POH layout of the player.